Fixed a bug.
Apparently, when the end of a CSV entry is just a bunch of NULL values, the python CSV reader will sometimes just not read in these values and return a list of smaller size than the number of the headers. This created an indexing error that is solved in this update.